v1.23.0

+++ nerve/converter/converter-lib/src/main/java/network/nerve/converter/heterogeneouschain/lib/helper/HtgERC20Helper.java
@@ -253,6 +253,9 @@ public class HtgERC20Helper implements BeanInitial {
         try {
             for (HeterogeneousAssetInfo info : assetInfos) {
                 HtgERC20Po erc20Po = htgERC20StorageService.findByAddress(info.getContractAddress());
+                if (erc20Po == null) {
+                    continue;
+                }
                 info.setAssetId(erc20Po.getAssetId());
                 htgERC20StorageService.deleteByAddress(erc20Po.getAddress());
                 successList.add(erc20Po);
